recipe ui refactor
