+++ app/src/main/java/com/jin/honey/feature/navigation/RootNavigation.kt
@@ -107,7 +107,7 @@ fun RootNavigation(
         ) {
             val menuName = it.arguments?.getString(Screens.MENU_MANE).orEmpty()
             val viewModel = remember { RecipeViewModel(GetRecipeUseCase(foodRepository)) }
-            RecipeScreen(viewModel, menuName)
+            RecipeScreen(viewModel, menuName) { navController.popBackStack() }
         }
     }
 }
