+++ app/src/main/java/com/jin/honey/feature/recipe/ui/content/RecipeHeader.kt
@@ -1,0 +1,41 @@
+package com.jin.honey.feature.recipe.ui.content
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.ArrowBackIosNew
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.sp
+import com.jin.honey.R
+
+@Composable
+fun RecipeHeader(onNavigateToBack: () -> Unit) {
+    Box(
+        modifier = Modifier.fillMaxWidth(),
+        contentAlignment = Alignment.CenterStart
+    ) {
+        IconButton(onNavigateToBack) {
+            Icon(
+                imageVector = Icons.Outlined.ArrowBackIosNew,
+                contentDescription = stringResource(R.string.ingredient_back_icon_desc),
+                tint = Color.Black
+            )
+        }
+        Text(
+            text = stringResource(R.string.recipe_title),
+            modifier = Modifier.fillMaxWidth(),
+            textAlign = TextAlign.Center,
+            fontSize = 20.sp,
+            fontWeight = FontWeight.Bold
+        )
+    }
+}
